Move api tests to their own module
